V1.13.0 (#200)
New Feature

#182 - Document Retry - documents that fail to process are added to a queue for reprocessing
#192 - Added POST /documents/{documentId}/actions/retry for retrying of failed document actions
#196 - Added OCR support for PDF Portfolio files
#194 - Added support for document content to be a "deep link"
#189 - Added document "soft delete"
#146 - Added insertedDate, completedDate to Document actions
#171 - Added Document Workflows
#179 - Added Cognito group that allows authentication but no authorization
#170 - Added Notification action
#168 - Added SMTP support to /configuration API
#169 - Added /groups API
#201 - Added version of FormKiQ that can be run via Docker

Updates

#166 - AWS Lambda end of support for Custom Runtime based on Amazon Linux 1
#188 - Intelligent document processing with OpenAI, uses Schemas to improve results
#119 - improved performance when removing documents

Bugs Fixes

#172 - Fixed Fulltext action ending up in a bad state
#185 - Fixed GET /documents/{documentId}/url returns full path of document instead of filename
#197 - POST /documents/compress is not available for users with READ ONLY access

+++ lambda-api/src/main/java/com/formkiq/stacks/api/handler/DocumentIdRequestHandler.java
@@ -27,6 +27,7 @@ import static com.formkiq.aws.dynamodb.SiteIdKeyGenerator.DEFAULT_SITE_ID;
 import static com.formkiq.aws.dynamodb.SiteIdKeyGenerator.createDatabaseKey;
 import static com.formkiq.aws.dynamodb.SiteIdKeyGenerator.createS3Key;
 import static com.formkiq.aws.dynamodb.objects.Objects.throwIfNull;
+import static com.formkiq.aws.dynamodb.objects.Strings.isEmpty;
 import static com.formkiq.aws.services.lambda.ApiResponseStatus.SC_CREATED;
 import static com.formkiq.aws.services.lambda.ApiResponseStatus.SC_NOT_FOUND;
 import static com.formkiq.aws.services.lambda.ApiResponseStatus.SC_OK;
@@ -52,6 +53,7 @@ import com.formkiq.aws.dynamodb.model.DocumentTag;
 import com.formkiq.aws.dynamodb.model.DynamicDocumentItem;
 import com.formkiq.aws.dynamodb.objects.DateUtil;
 import com.formkiq.aws.s3.S3ObjectMetadata;
+import com.formkiq.aws.s3.S3PresignerService;
 import com.formkiq.aws.s3.S3Service;
 import com.formkiq.aws.services.lambda.ApiAuthorization;
 import com.formkiq.aws.services.lambda.ApiGatewayRequestEvent;
@@ -190,27 +192,31 @@ public class DocumentIdRequestHandler
     String siteId = authorization.siteId();
     String documentId = event.getPathParameters().get("documentId");
 
-    logger.log("deleting object " + documentId + " from bucket '" + documentBucket + "'");
+    if (awsservice.debug()) {
+      logger.log("deleting object " + documentId + " from bucket '" + documentBucket + "'");
+    }
 
     DocumentService service = awsservice.getExtension(DocumentService.class);
     DocumentItem item = service.findDocument(siteId, documentId);
     throwIfNull(item, new DocumentNotFoundException(documentId));
 
+    boolean softDelete = "true".equals(event.getQueryStringParameter("softDelete"));
+
     try {
 
-      S3Service s3Service = awsservice.getExtension(S3Service.class);
+      if (!softDelete) {
+        S3Service s3Service = awsservice.getExtension(S3Service.class);
 
-      String s3Key = SiteIdKeyGenerator.createS3Key(siteId, documentId);
-      S3ObjectMetadata md = s3Service.getObjectMetadata(documentBucket, s3Key, null);
+        String s3Key = SiteIdKeyGenerator.createS3Key(siteId, documentId);
+        S3ObjectMetadata md = s3Service.getObjectMetadata(documentBucket, s3Key, null);
 
-      if (md.isObjectExists()) {
-        s3Service.deleteObject(documentBucket, s3Key, null);
-
-      } else {
-
-        if (!service.deleteDocument(siteId, documentId)) {
-          throw new NotFoundException("Document " + documentId + " not found.");
+        if (md.isObjectExists()) {
+          s3Service.deleteObject(documentBucket, s3Key, null);
         }
+      }
+
+      if (!service.deleteDocument(siteId, documentId, softDelete)) {
+        throw new NotFoundException("Document " + documentId + " not found.");
       }
 
       ApiResponse resp = new ApiMessageResponse("'" + documentId + "' object deleted");
@@ -241,7 +247,7 @@ public class DocumentIdRequestHandler
     if (documentId != null) {
       Duration duration = Duration.ofHours(DEFAULT_DURATION_HOURS);
       String key = createS3Key(siteId, documentId);
-      S3Service s3Service = awsservice.getExtension(S3Service.class);
+      S3PresignerService s3Service = awsservice.getExtension(S3PresignerService.class);
 
       Map<String, String> map = Map.of("checksum", UUID.randomUUID().toString());
       url = s3Service.presignPutUrl(awsservice.environment("DOCUMENTS_S3_BUCKET"), key, duration,
@@ -374,7 +380,7 @@ public class DocumentIdRequestHandler
 
       validateTagSchema(awsservice, siteId, item, item.getUserId(), isUpdate);
       validateTags(item);
-      validateActions(awsservice, siteId, item);
+      validateActions(awsservice, siteId, item, authorization);
 
       putObjectToStaging(logger, awsservice, maxDocumentCount, siteId, item);
 
@@ -434,10 +440,14 @@ public class DocumentIdRequestHandler
   }
 
   private void validateActions(final AwsServiceCache awsservice, final String siteId,
-      final DynamicDocumentItem item) throws ValidationException {
+      final DynamicDocumentItem item, final ApiAuthorization authorization)
+      throws ValidationException {
 
     List<DynamicObject> objs = item.getList("actions");
     if (!objs.isEmpty()) {
+
+      objs.stream().forEach(a -> a.put("userId", authorization.username()));
+      item.put("actions", objs);
 
       ConfigService configsService = awsservice.getExtension(ConfigService.class);
       DynamicObject configs = configsService.get(siteId);
@@ -457,7 +467,7 @@ public class DocumentIdRequestHandler
                 .collect(Collectors.toMap(e -> e.getKey(), e -> e.getValue().toString()))
             : Collections.emptyMap();
 
-        return new Action().type(type).parameters(parameters);
+        return new Action().type(type).parameters(parameters).userId(o.getString("userId"));
       }).collect(Collectors.toList());
 
       for (Action action : actions) {
@@ -503,7 +513,8 @@ public class DocumentIdRequestHandler
 
     boolean isFolder = isFolder(item);
 
-    if (!isFolder && !item.hasString("content") && item.getList("documents").isEmpty()) {
+    if (!isFolder && !item.hasString("content") && item.getList("documents").isEmpty()
+        && isEmpty(item.getDeepLinkPath())) {
       throw new BadException("Invalid JSON body.");
     }
